chore(sdn-controller): Put TLS socket code in its own file
