+++ packages/xo-server-sdn-controller/src/index.js
@@ -10,6 +10,7 @@ import { join } from 'path'
 
 import { OvsdbClient } from './protocol/ovsdb-client'
 import { PrivateNetwork } from './private-network/private-network'
+import { TlsHelper } from './utils/tls-helper'
 
 // =============================================================================
 
@@ -335,6 +336,8 @@ class SDNController extends EventEmitter {
     this._prevVni = 0
 
     this.ovsdbClients = {}
+
+    this._tlsHelper = new TlsHelper()
   }
 
   // ---------------------------------------------------------------------------
@@ -369,10 +372,11 @@ class SDNController extends EventEmitter {
       fileRead(join(certDirectory, CLIENT_CERT)),
       fileRead(join(certDirectory, CA_CERT)),
     ])
-
-    forOwn(this.ovsdbClients, client => {
-      client.updateCertificates(this._clientKey, this._clientCert, this._caCert)
-    })
+    this._tlsHelper.updateCertificates(
+      this._clientKey,
+      this._clientCert,
+      this._caCert
+    )
     const updatedPools = []
     await Promise.all(
       map(this.privateNetworks, async privateNetworks => {
@@ -1120,12 +1124,7 @@ class SDNController extends EventEmitter {
       return
     }
 
-    const client = new OvsdbClient(
-      host,
-      this._clientKey,
-      this._clientCert,
-      this._caCert
-    )
+    const client = new OvsdbClient(host, this._tlsHelper)
     this.ovsdbClients[host.$ref] = client
   }
 }
